FEAT MemberService REFACTOR AuthController
